Have all the evtest examples use the same args/prompt code
